Adaptor manipulations moved from builders to media classes. Background hirarchy extended with 2 subclasses (OneColorBackground and ImageBackground). Adaptors and other dependencies are not set from external functions now, they should be overrided in test classes. Image align, valign and scaletype moved to ImageProperties structure. BackgroundTest and BackgroundBuilderTests refactored. Added media general tests.

+++ player/media/AudioHandler.hpp
@@ -16,7 +16,7 @@ public:
     void play() override;
     void stop() override;
     void connect(OnAudioFinished callback) override;
-    void setVolume(double volume) override;
+    void setVolume(int volume) override;
 
 private:
     bool busMessageWatch(const Gst::RefPtr<Gst::Message>& message);
