+++ player/media/IVideoHandler.hpp
@@ -22,5 +22,5 @@ public:
     virtual void play() = 0;
     virtual void stop() = 0;
     virtual void connect(OnVideoFinished callback) = 0;
-    virtual void setVolume(double volume) = 0;
+    virtual void setVolume(int volume) = 0;
 };
